+++ src/index.ts
@@ -1,2 +1,5 @@
 export * from './main';
 export * from './runner';
+import { Patch } from 'awscdk-81-patch';
+
+Patch.apply();