+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.16](https://github.com/pahud/cdk-gitlab/compare/v0.1.15...v0.1.16) (2021-01-02)
+
+### 0.1.15 (2020-11-13)
+
+
+### Bug Fixes
+
+* helmRunnerOptions to optional ([#26](https://github.com/pahud/cdk-gitlab/issues/26)) ([e5fe258](https://github.com/pahud/cdk-gitlab/commit/e5fe2588b9f1ab680889a91b95187c77937d5e5a)), closes [#25](https://github.com/pahud/cdk-gitlab/issues/25)
+
+### 0.1.14 (2020-11-13)
+
+
+### Features
+
+* support helm chart runner ([#23](https://github.com/pahud/cdk-gitlab/issues/23)) ([ee04fdc](https://github.com/pahud/cdk-gitlab/commit/ee04fdcfaefa39ab8f143096ac60a938a3e06397)), closes [#24](https://github.com/pahud/cdk-gitlab/issues/24)
+
+### 0.1.13 (2020-11-11)
+
+### 0.1.12 (2020-11-08)
+
+### 0.1.11 (2020-11-08)
+
+### 0.1.10 (2020-11-08)
+
+### 0.1.9 (2020-11-08)
+
+### 0.1.8 (2020-11-07)
+
+### 0.1.7 (2020-11-06)
+
 ### 0.1.6 (2020-11-06)
 
 ### 0.1.5 (2020-11-05)